services/scraper/pyproject.toml
[project]
name = "news_scraper"
version = "0.1.0"
requires-python = ">=3.12"
description = "Ingestion service — YouTube RSS, blog RSS, web search"
dependencies = [
    "fastapi>=0.136.1",
    "uvicorn[standard]>=0.32",
    "typer>=0.24.2",
    "feedparser>=6.0.11",
    "youtube-transcript-api>=0.6.2",
    "python-dateutil>=2.9",
    "openai-agents>=0.14.5",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.5",
    "loguru>=0.7.3",
    "news_schemas",
    "news_config",
    "news_observability",
    "news_db",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/news_scraper"]
